requirements.txt
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
jsonpickle==3.0.1
mypy-extensions==1.0.0
python-decouple==3.8
python-digitalocean==1.17.0
python-dotenv==1.0.0
python-environ==0.4.54
requests==2.28.2
typing-extensions==4.5.0
urllib3==1.26.15
